+++ e2e/commands/export.e2e.js
@@ -177,7 +177,7 @@ describe('bit export command', function () {
     });
   });
 
-  describe('after import', () => {
+  describe('imported (v1), exported (v2) and then exported again (v3)', () => {
     before(() => {
       helper.reInitLocalScope();
       helper.reInitRemoteScope();
@@ -185,20 +185,23 @@ describe('bit export command', function () {
       helper.createComponentBarFoo();
       helper.addComponentBarFoo();
       helper.commitComponentBarFoo();
-      helper.exportComponent('bar/foo');
+      helper.exportComponent('bar/foo'); // v1
 
       helper.reInitLocalScope();
       helper.addRemoteScope();
       helper.importComponent('bar/foo');
 
+      helper.createFile(path.join('components', 'bar', 'foo', 'bar'), 'foo.js', 'console.log("got foo v2")');
       helper.commitComponentBarFoo();
-      helper.exportComponent(`${helper.remoteScope}/bar/foo`);
+      helper.exportComponent(`${helper.remoteScope}/bar/foo`); // v2
+
+      helper.createFile(path.join('components', 'bar', 'foo', 'bar'), 'foo.js', 'console.log("got foo v3")');
+      helper.commitComponentBarFoo();
+      helper.exportComponent(`${helper.remoteScope}/bar/foo`); // v3
     });
     it('should export it with no errors', () => {
-      const output = helper.runCmd(`bit list ${helper.remoteScope}`);
-      expect(output.includes('Total 1 components')).to.be.true;
-      expect(output.includes('bar/foo')).to.be.true;
-      expect(output.includes('2')).to.be.true; // this is the version
+      const output = helper.listRemoteScope();
+      expect(output.includes(`${helper.remoteScope}/bar/foo@3`)).to.be.true;
     });
   });
 
@@ -274,6 +277,32 @@ describe('bit export command', function () {
 
       helper.commitComponentBarFoo(); // v3
       helper.exportComponent(`${helper.remoteScope}/bar/foo`);
+    });
+    it('should export it with no errors', () => {
+      const output = helper.listRemoteScope();
+      expect(output.includes('Total 1 components')).to.be.true;
+      expect(output.includes('bar/foo')).to.be.true;
+      expect(output.includes('3')).to.be.true; // this is the version
+    });
+  });
+
+  describe('exporting version 3 of a component as an author', () => {
+    before(() => {
+      helper.reInitLocalScope();
+      helper.reInitRemoteScope();
+      helper.addRemoteScope();
+      helper.createComponent('bar', 'foo.js', 'console.log("got foo v1")');
+      helper.addComponentBarFoo();
+      helper.commitComponentBarFoo(); // v1
+      helper.exportComponent('bar/foo');
+
+      helper.createComponent('bar', 'foo.js', 'console.log("got foo v2")');
+      helper.commitComponentBarFoo(); // v2
+      helper.exportComponent('bar/foo');
+
+      helper.createComponent('bar', 'foo.js', 'console.log("got foo v3")');
+      helper.commitComponentBarFoo(); // v3
+      helper.exportComponent('bar/foo');
     });
     it('should export it with no errors', () => {
       const output = helper.listRemoteScope();