+++ src/consumer/bit-map/bit-map.js
@@ -4,7 +4,7 @@ import R from 'ramda';
 import find from 'lodash.find';
 import pickBy from 'lodash.pickby';
 import logger from '../../logger/logger';
-import { BIT_MAP, DEFAULT_INDEX_NAME, BIT_JSON, COMPONENT_ORIGINS, DEPENDENCIES_DIR } from '../../constants';
+import { BIT_MAP, DEFAULT_INDEX_NAME, DEFAULT_INDEX_TS_NAME, BIT_JSON, COMPONENT_ORIGINS, DEPENDENCIES_DIR } from '../../constants';
 import { InvalidBitMap, MissingMainFile, MissingBitMapComponent } from './exceptions';
 import { BitId } from '../../bit-id';
 import { readFile, outputFile } from '../../utils';
@@ -89,24 +89,32 @@ export default class BitMap {
     });
   }
 
+  // todo - need to move to bit-javascript
+  _searchMainFile(baseMainFile: string, files: ComponentMapFile[]) {
+    let newBaseMainFile;
+    // Search the relativePath of the main file
+    let mainFileFromFiles = R.find(R.propEq('relativePath', baseMainFile))(files);
+    // Search the base name of the main file and transfer to relativePath
+    if (!mainFileFromFiles) {
+      mainFileFromFiles = R.find(R.propEq('name', baseMainFile))(files);
+      newBaseMainFile = mainFileFromFiles ? mainFileFromFiles.relativePath : baseMainFile;
+    }
+    return { mainFileFromFiles, baseMainFile: newBaseMainFile || baseMainFile };
+  }
   _getMainFile(mainFile: string, componentMap: ComponentMap) {
-    let baseMainFile = mainFile || DEFAULT_INDEX_NAME;
     const files = componentMap.files.filter(file => !file.test);
     // Take the file path as main in case there is only one file
     if (!mainFile && files.length === 1) return files[0].relativePath;
 
-    // Search the relativePath of the main file
-    let mainFileFromFiles = R.find(R.propEq('relativePath', baseMainFile))(files);
+    // search main file (index.js or index.ts in case no ain file was entered - move to bit-javascript
+    let searchResult = this._searchMainFile(mainFile, files)
+    if (!searchResult.mainFileFromFiles) searchResult = this._searchMainFile(DEFAULT_INDEX_NAME, files)
+    if (!searchResult.mainFileFromFiles) searchResult = this._searchMainFile(DEFAULT_INDEX_TS_NAME, files);
 
-    // Search the base name of the main file and transfer to relativePath
-    if (!mainFileFromFiles) {
-      mainFileFromFiles = R.find(R.propEq('name', baseMainFile))(files);
-      baseMainFile = mainFileFromFiles ? mainFileFromFiles.relativePath : baseMainFile;
-    }
 
     // When there is more then one file and the main file not found there
-    if (!mainFileFromFiles) throw new MissingMainFile(baseMainFile, files.map((file) => file.relativePath));
-    return baseMainFile;
+    if (!searchResult.mainFileFromFiles) throw new MissingMainFile(searchResult.baseMainFile, files.map((file) => file.relativePath));
+    return searchResult.baseMainFile;
   }
 
   addDependencyToParent(parent: BitId, dependency: string): void {
@@ -208,8 +216,8 @@ export default class BitMap {
     const newIdString = id.toString();
     const olderComponentsIds = Object.keys(this.components).filter(componentId => BitId
       .parse(componentId).toStringWithoutScopeAndVersion() === id.toStringWithoutScopeAndVersion()
-      && componentId !== newIdString
-      && this.components[componentId].origin !== COMPONENT_ORIGINS.NESTED);
+    && componentId !== newIdString
+    && this.components[componentId].origin !== COMPONENT_ORIGINS.NESTED);
 
     if (!olderComponentsIds.length) {
       logger.debug(`bit-map: no need to update ${newIdString}`);