+++ src/environment/environment.js
@@ -4,18 +4,18 @@ import fs from 'fs-extra';
 import path from 'path';
 import R from 'ramda';
 import npmClient from '../npm-client';
-import { Scope } from '../scope';
-import { flattenDependencies } from '../scope/flatten-dependencies';
+import { Scope, ComponentWithDependencies } from '../scope';
 import { BitId } from '../bit-id';
-import { Component } from '../consumer/component/consumer-component';
 import { BITS_DIRNAME, ISOLATED_ENV_ROOT } from '../constants';
-import { Driver } from '../driver';
 import { mkdirp } from '../utils';
 import logger from '../logger/logger';
+import { Consumer } from '../consumer';
+import Component from '../consumer/component';
 
 export default class Environment {
   path: string;
   scope: Scope;
+  consumer: Consumer;
 
   constructor(scope: Scope) {
     this.scope = scope;
@@ -23,53 +23,16 @@ export default class Environment {
     logger.debug(`creating a new isolated environment at ${this.path}`);
   }
 
-  create(): Promise<> {
-    return new Promise((resolve, reject) => {
-      mkdirp(this.path, (err) => {
-        if (err) return reject(err);
-        return resolve();
-      });
-    });
-  }
-
-  importComponent(rawId: string): Promise<Component> {
-    const bitId = BitId.parse(rawId);
-    return this.scope.get(bitId)
-      .then((component) => {
-        return this._writeToEnvironmentDir([component.component])
-          .then(components => components[0]);
-      });
-  }
-
-  _writeToEnvironmentDir(components: Component[]): Promise<Component[]> {
-    return Promise.all(components.map((component) => {
-      const bitPath = this.getComponentPath(component);
-      return component.write({ bitDir: bitPath });
-    }));
-  }
-
-  importDependencies(component: Component): Promise<Component[]> {
-    if (!component.dependencies || !component.dependencies.length) return Promise.resolve([]);
-    return this.scope.getMany(component.dependencies).then((componentDependenciesArr) => {
-      const components = flattenDependencies(componentDependenciesArr);
-      return this._writeToEnvironmentDir(components);
-    });
+  async create(): Promise<> {
+    await mkdirp(this.path);
+    this.consumer = await Consumer.createWithExistingScope(this.path, this.scope);
   }
 
   installNpmPackages(components: Component[]): Promise<*> {
     return Promise.all(components.map((component) => {
       if (R.isEmpty(component.packageDependencies)) return Promise.resolve();
-      const componentPath = this.getComponentPath(component);
-      return npmClient.install(component.packageDependencies, { cwd: componentPath });
+      return npmClient.install(component.packageDependencies, { cwd: component.writtenPath });
     }));
-  }
-
-  bindFromDriver(component: Component) {
-    const driver = Driver.load(component.lang).getDriver(false);
-    if (driver) {
-      return driver.bind({ projectRoot: this.path });
-    }
-    return Promise.resolve();
   }
 
   /**
@@ -79,19 +42,15 @@ export default class Environment {
    * @param rawId
    * @return {Promise.<Component>}
    */
-  importE2E(rawId: string): Promise<Component> {
-    return this.importComponent(rawId).then((component) => {
-      return this.importDependencies(component).then((dependencies) => {
-        const components = dependencies.concat(component);
-        return this.installNpmPackages(components)
-          .then(() => this.bindFromDriver(component))
-          .then(() => component);
-      });
-    });
-  }
-
-  getComponentPath(component: Component): string {
-    return path.join(this.path, BITS_DIRNAME, component.id.toFullPath());
+  async importE2E(rawId: string): Promise<Component> {
+    const bitId = BitId.parse(rawId);
+    const componentDependenciesArr = await this.scope.getMany([bitId]);
+    await this.consumer.writeToComponentsDir(componentDependenciesArr);
+    const componentWithDependencies: ComponentWithDependencies = R.head(componentDependenciesArr);
+    const componentWithDependenciesFlatten = [componentWithDependencies.component,
+      ...componentWithDependencies.dependencies];
+    await this.installNpmPackages(componentWithDependenciesFlatten);
+    return componentWithDependencies.component;
   }
 
   getPath(): string {